+++ packages/bpk-foundations-web/tokens/base.es6.d.ts
@@ -1,0 +1,845 @@
+/*
+ * 
+ * Backpack - Skyscanner's Design System
+ * 
+ * Copyright 2016-2021 Skyscanner Ltd
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * 
+ */
+export declare const colorSkyBlueShade03 = "rgb(2, 18, 44)" as const;
+export declare const colorPrimaryGradientLight = "rgb(0, 98, 227)" as const;
+export declare const colorErfoud = "rgb(255, 181, 77)" as const;
+export declare const colorValensole = "rgb(165, 155, 200)" as const;
+export declare const colorMonteverde = "rgb(0, 166, 152)" as const;
+export declare const colorSkyGrayTint01 = "rgb(68, 69, 96)" as const;
+export declare const colorBlack = "rgb(1, 9, 19)" as const;
+export declare const colorSkyGrayTint02 = "rgb(104, 105, 127)" as const;
+export declare const colorSagano = "rgb(208, 238, 236)" as const;
+export declare const colorSkyGrayTint03 = "rgb(143, 144, 160)" as const;
+export declare const colorSkyGrayTint04 = "rgb(178, 178, 191)" as const;
+export declare const colorHarbour = "rgb(246, 221, 225)" as const;
+export declare const colorSkyGrayTint05 = "rgb(205, 205, 215)" as const;
+export declare const colorSkyGrayTint06 = "rgb(221, 221, 229)" as const;
+export declare const colorBlackTint01 = "rgb(29, 27, 32)" as const;
+export declare const colorSkyGrayTint07 = "rgb(241, 242, 248)" as const;
+export declare const colorBlackTint02 = "rgb(44, 44, 46)" as const;
+export declare const colorAbisko = "rgb(90, 72, 155)" as const;
+export declare const colorBlackTint03 = "rgb(58, 58, 60)" as const;
+export declare const colorBlackTint04 = "rgb(72, 72, 74)" as const;
+export declare const colorWhite = "rgb(255, 255, 255)" as const;
+export declare const colorBlackTint05 = "rgb(99, 99, 102)" as const;
+export declare const colorBlackTint06 = "rgb(142, 142, 147)" as const;
+export declare const colorPanjin = "rgb(209, 67, 91)" as const;
+export declare const colorKolkata = "rgb(255, 148, 0)" as const;
+export declare const colorGlencoe = "rgb(115, 206, 198)" as const;
+export declare const colorTochigi = "rgb(225, 221, 236)" as const;
+export declare const colorPetra = "rgb(255, 171, 149)" as const;
+export declare const colorSkyBlue = "rgb(0, 98, 227)" as const;
+export declare const colorBagan = "rgb(255, 235, 208)" as const;
+export declare const colorHillier = "rgb(225, 139, 150)" as const;
+export declare const colorSkyBlueTint01 = "rgb(109, 159, 235)" as const;
+export declare const colorSkyBlueTint02 = "rgb(157, 192, 242)" as const;
+export declare const colorBunol = "rgb(255, 123, 89)" as const;
+export declare const colorSkyBlueTint03 = "rgb(205, 223, 248)" as const;
+export declare const colorSkyGray = "rgb(17, 18, 54)" as const;
+export declare const colorNara = "rgb(255, 231, 224)" as const;
+export declare const colorSkyBlueShade01 = "rgb(8, 78, 178)" as const;
+export declare const colorSkyBlueShade02 = "rgb(4, 39, 89)" as const;
+export declare const privateBadgeBackgroundNormalDay = "rgb(239, 243, 248)" as const;
+export declare const privateBadgeBackgroundNormalNight = "rgb(36, 51, 70)" as const;
+export declare const canvasDay = "rgb(255, 255, 255)" as const;
+export declare const canvasNight = "rgb(1, 9, 19)" as const;
+export declare const canvasContrastDay = "rgb(239, 243, 248)" as const;
+export declare const canvasContrastNight = "rgb(1, 9, 19)" as const;
+export declare const coreEcoDay = "rgb(15, 161, 169)" as const;
+export declare const coreEcoNight = "rgb(15, 161, 169)" as const;
+export declare const coreAccentDay = "rgb(0, 98, 227)" as const;
+export declare const coreAccentNight = "rgb(132, 233, 255)" as const;
+export declare const corePrimaryDay = "rgb(5, 32, 60)" as const;
+export declare const corePrimaryNight = "rgb(5, 65, 132)" as const;
+export declare const statusSuccessSpotNight = "rgb(98, 241, 198)" as const;
+export declare const statusSuccessFillNight = "rgb(177, 255, 231)" as const;
+export declare const statusSuccessSpotDay = "rgb(12, 131, 138)" as const;
+export declare const statusSuccessFillDay = "rgb(212, 255, 242)" as const;
+export declare const statusDangerSpotNight = "rgb(255, 100, 156)" as const;
+export declare const statusWarningSpotNight = "rgb(254, 235, 135)" as const;
+export declare const statusDangerFillNight = "rgb(255, 202, 221)" as const;
+export declare const statusWarningFillNight = "rgb(251, 241, 187)" as const;
+export declare const statusDangerSpotDay = "rgb(231, 8, 102)" as const;
+export declare const statusWarningSpotDay = "rgb(245, 93, 66)" as const;
+export declare const statusDangerFillDay = "rgb(255, 233, 249)" as const;
+export declare const statusWarningFillDay = "rgb(255, 247, 207)" as const;
+export declare const surfaceHeroNight = "rgb(1, 9, 19)" as const;
+export declare const surfaceDefaultNight = "rgb(19, 29, 43)" as const;
+export declare const surfaceHeroDay = "rgb(0, 98, 227)" as const;
+export declare const surfaceDefaultDay = "rgb(255, 255, 255)" as const;
+export declare const surfaceContrastNight = "rgb(1, 9, 19)" as const;
+export declare const surfaceContrastDay = "rgb(5, 32, 60)" as const;
+export declare const surfaceElevatedNight = "rgb(36, 51, 70)" as const;
+export declare const surfaceSubtleNight = "rgb(36, 51, 70)" as const;
+export declare const surfaceElevatedDay = "rgb(255, 255, 255)" as const;
+export declare const surfaceHighlightNight = "rgb(36, 51, 70)" as const;
+export declare const surfaceSubtleDay = "rgb(227, 240, 255)" as const;
+export declare const surfaceHighlightDay = "rgb(224, 228, 233)" as const;
+export declare const textOnDarkDay = "rgb(255, 255, 255)" as const;
+export declare const textOnLightNight = "rgb(1, 9, 19)" as const;
+export declare const textDisabledOnDarkDay = "rgba(255, 255, 255, 0.5)" as const;
+export declare const textOnLightDay = "rgb(22, 22, 22)" as const;
+export declare const textLinkNight = "rgb(132, 233, 255)" as const;
+export declare const textHeroNight = "rgb(1, 9, 19)" as const;
+export declare const textDisabledNight = "rgba(255, 255, 255, 0.2)" as const;
+export declare const textPrimaryDarkColor = "rgb(255, 255, 255)" as const;
+export declare const textLinkDay = "rgb(0, 98, 227)" as const;
+export declare const textHeroDay = "rgb(0, 98, 227)" as const;
+export declare const textQuaternaryDarkColor = "rgb(142, 142, 147)" as const;
+export declare const textPrimaryNight = "rgb(255, 255, 255)" as const;
+export declare const textQuaternaryLightColor = "rgb(143, 144, 160)" as const;
+export declare const textDisabledDay = "rgba(0, 0, 0, 0.2)" as const;
+export declare const textPrimaryLightColor = "rgb(17, 18, 54)" as const;
+export declare const textPrimaryDay = "rgb(22, 22, 22)" as const;
+export declare const textErrorNight = "rgb(255, 100, 156)" as const;
+export declare const textErrorDay = "rgb(231, 8, 102)" as const;
+export declare const textTertiaryDarkColor = "rgb(142, 142, 147)" as const;
+export declare const textSecondaryDarkColor = "rgb(178, 178, 191)" as const;
+export declare const textPrimaryInverseNight = "rgb(1, 9, 19)" as const;
+export declare const textSecondaryNight = "rgb(189, 196, 203)" as const;
+export declare const textOnDarkNight = "rgb(255, 255, 255)" as const;
+export declare const textSecondaryLightColor = "rgb(104, 105, 127)" as const;
+export declare const textPrimaryInverseDay = "rgb(255, 255, 255)" as const;
+export declare const textDisabledOnDarkNight = "rgba(255, 255, 255, 0.5)" as const;
+export declare const textTertiaryLightColor = "rgb(143, 144, 160)" as const;
+export declare const textSecondaryDay = "rgb(98, 105, 113)" as const;
+export declare const scrimDay = "rgba(0, 0, 0, 0.7)" as const;
+export declare const backgroundElevation03DarkColor = "rgb(58, 58, 60)" as const;
+export declare const backgroundElevation02DarkColor = "rgb(44, 44, 46)" as const;
+export declare const lineOnDarkNight = "rgb(68, 80, 95)" as const;
+export declare const lineDarkColor = "rgb(72, 72, 74)" as const;
+export declare const backgroundElevation01DarkColor = "rgb(29, 27, 32)" as const;
+export declare const lineOnDarkDay = "rgba(255, 255, 255, 0.5)" as const;
+export declare const backgroundAlternativeSecondaryDarkColor = "rgb(29, 27, 32)" as const;
+export declare const backgroundLightColor = "rgb(255, 255, 255)" as const;
+export declare const backgroundDarkColor = "rgb(0, 0, 0)" as const;
+export declare const backgroundAlternativeSecondaryLightColor = "rgb(255, 255, 255)" as const;
+export declare const lineNight = "rgb(68, 80, 95)" as const;
+export declare const colorSystemGreen = "rgb(0, 166, 152)" as const;
+export declare const backgroundElevation01LightColor = "rgb(255, 255, 255)" as const;
+export declare const lineLightColor = "rgb(205, 205, 215)" as const;
+export declare const backgroundElevation02LightColor = "rgb(255, 255, 255)" as const;
+export declare const lineDay = "rgb(193, 199, 207)" as const;
+export declare const backgroundElevation03LightColor = "rgb(255, 255, 255)" as const;
+export declare const backgroundTertiaryDarkColor = "rgb(44, 44, 46)" as const;
+export declare const backgroundSecondaryDarkColor = "rgb(29, 27, 32)" as const;
+export declare const backgroundAlternativeLightColor = "rgb(241, 242, 248)" as const;
+export declare const primaryDarkColor = "rgb(109, 159, 235)" as const;
+export declare const colorSystemRed = "rgb(209, 67, 91)" as const;
+export declare const backgroundAlternativeDarkColor = "rgb(0, 0, 0)" as const;
+export declare const primaryLightColor = "rgb(0, 98, 227)" as const;
+export declare const backgroundTertiaryLightColor = "rgb(255, 255, 255)" as const;
+export declare const backgroundSecondaryLightColor = "rgb(241, 242, 248)" as const;
+export declare const scrimNight = "rgba(0, 0, 0, 0.7)" as const;
+export declare const privateInfoBannerDefaultDay = "rgb(239, 243, 248)" as const;
+export declare const privateInfoBannerOnContrastDay = "rgb(255, 255, 255)" as const;
+export declare const privateInfoBannerSuccessDay = "rgb(12, 131, 138)" as const;
+export declare const privateInfoBannerErrorNight = "rgb(255, 100, 156)" as const;
+export declare const privateInfoBannerInfoNight = "rgb(189, 196, 203)" as const;
+export declare const privateInfoBannerErrorDay = "rgb(231, 8, 102)" as const;
+export declare const privateInfoBannerInfoDay = "rgb(98, 105, 113)" as const;
+export declare const privateInfoBannerWarningNight = "rgb(254, 235, 135)" as const;
+export declare const privateInfoBannerWarningDay = "rgb(245, 93, 66)" as const;
+export declare const privateInfoBannerDefaultNight = "rgb(19, 29, 43)" as const;
+export declare const privateInfoBannerOnContrastNight = "rgb(19, 29, 43)" as const;
+export declare const privateInfoBannerSuccessNight = "rgb(98, 241, 198)" as const;
+export declare const marcommsBerry = "rgb(231, 8, 102)" as const;
+export declare const marcommsCharcoal = "rgb(22, 22, 22)" as const;
+export declare const marcommsOrange = "rgb(255, 123, 89)" as const;
+export declare const marcommsGrey40 = "rgb(98, 105, 113)" as const;
+export declare const marcommsBlueBright = "rgb(161, 238, 255)" as const;
+export declare const marcommsYellow = "rgb(254, 235, 135)" as const;
+export declare const marcommsGreenBright = "rgb(147, 255, 222)" as const;
+export declare const marcommsPink = "rgb(255, 163, 229)" as const;
+export declare const marcommsSkyBlue = "rgb(0, 98, 227)" as const;
+export declare const marcommsDarkSky = "rgb(5, 32, 60)" as const;
+export declare const marcommsPurple = "rgb(142, 71, 186)" as const;
+export declare const marcommsWhite = "rgb(255, 255, 255)" as const;
+export declare const marcommsYellowMuted = "rgb(255, 247, 207)" as const;
+export declare const marcommsPinkMuted = "rgb(255, 233, 249)" as const;
+export declare const marcommsGreenMuted = "rgb(212, 255, 242)" as const;
+export declare const marcommsBlueMuted = "rgb(217, 248, 255)" as const;
+export declare const marcommsEcoGreen = "rgb(15, 161, 169)" as const;
+export declare const privateChipOnDarkPressedStrokeDay = "rgb(255, 255, 255)" as const;
+export declare const privateChipOnDarkPressedStrokeNight = "rgb(5, 65, 132)" as const;
+export declare const privateChipOnDarkOnBackgroundDay = "rgb(255, 255, 255)" as const;
+export declare const privateChipOnDarkOnBackgroundNight = "rgb(5, 65, 132)" as const;
+export declare const privateChipDisabledBackgroundDay = "rgb(224, 228, 233)" as const;
+export declare const privateChipDisabledBackgroundNight = "rgb(11, 18, 29)" as const;
+export declare const privateChipOnDarkOnDismissIconDay = "rgb(98, 105, 113)" as const;
+export declare const privateChipOnDarkOnDismissIconNight = "rgba(255, 255, 255, 0.5)" as const;
+export declare const privateMapPreviousSelectionDay = "rgb(207, 228, 255)" as const;
+export declare const privateMapClusterPinNight = "rgb(255, 255, 255)" as const;
+export declare const privateMapClusterPinPreviousSelectionNight = "rgb(148, 195, 255)" as const;
+export declare const privateMapClusterPinDay = "rgb(5, 32, 60)" as const;
+export declare const privateMapClusterPinPreviousSelectionDay = "rgb(148, 195, 255)" as const;
+export declare const privateMapMarkerViewedForegroundNight = "rgba(0, 0, 0, 0.8)" as const;
+export declare const privateMapMarkerViewedForegroundDay = "rgba(255, 255, 255, 0.8)" as const;
+export declare const privateMapPoiPinNight = "rgb(142, 71, 186)" as const;
+export declare const privateMapPreviousSelectionNight = "rgb(207, 228, 255)" as const;
+export declare const privateMapPoiPinDay = "rgb(142, 71, 186)" as const;
+export declare const privateButtonSecondaryPressedBackgroundNight = "rgb(1, 9, 19)" as const;
+export declare const privateButtonLinkNormalForegroundNight = "rgb(132, 233, 255)" as const;
+export declare const privateButtonLinkOnDarkDisabledForegroundNight = "rgba(255, 255, 255, 0.2)" as const;
+export declare const privateButtonSecondaryNormalBackgroundNight = "rgb(36, 51, 70)" as const;
+export declare const privateButtonPrimaryOnLightPressedBackgroundDay = "rgb(21, 70, 121)" as const;
+export declare const privateButtonFeaturedNormalBackgroundDay = "rgb(0, 98, 227)" as const;
+export declare const privateButtonFeaturedPressedBackgroundNight = "rgb(209, 247, 255)" as const;
+export declare const privateButtonPrimaryOnLightNormalBackgroundNight = "rgb(2, 77, 175)" as const;
+export declare const privateButtonSecondaryPressedBackgroundDay = "rgb(193, 199, 207)" as const;
+export declare const privateButtonLinkNormalForegroundDay = "rgb(0, 98, 227)" as const;
+export declare const privateButtonPrimaryOnLightDisabledForegroundNight = "rgba(255, 255, 255, 0.2)" as const;
+export declare const privateButtonSecondaryOnDarkDisabledBackgroundNight = "rgb(11, 18, 29)" as const;
+export declare const privateButtonDestructiveNormalForegroundNight = "rgb(255, 100, 156)" as const;
+export declare const privateButtonLinkOnDarkDisabledForegroundDay = "rgba(255, 255, 255, 0.2)" as const;
+export declare const privateButtonSecondaryNormalBackgroundDay = "rgb(224, 228, 233)" as const;
+export declare const privateButtonPrimaryOnLightNormalBackgroundDay = "rgb(5, 32, 60)" as const;
+export declare const privateButtonPrimaryOnLightDisabledForegroundDay = "rgba(0, 0, 0, 0.2)" as const;
+export declare const privateButtonPrimaryOnDarkNormalBackgroundNight = "rgb(255, 255, 255)" as const;
+export declare const privateButtonFeaturedPressedBackgroundDay = "rgb(2, 77, 175)" as const;
+export declare const privateButtonPrimaryOnDarkDisabledForegroundNight = "rgba(255, 255, 255, 0.2)" as const;
+export declare const privateButtonSecondaryOnDarkDisabledBackgroundDay = "rgb(11, 18, 29)" as const;
+export declare const privateButtonSecondaryOnDarkPressedBackgroundNight = "rgb(1, 9, 19)" as const;
+export declare const privateButtonDestructiveNormalForegroundDay = "rgb(231, 8, 102)" as const;
+export declare const privateButtonPrimaryOnDarkNormalBackgroundDay = "rgb(255, 255, 255)" as const;
+export declare const privateButtonPrimaryOnDarkDisabledForegroundDay = "rgba(0, 0, 0, 0.2)" as const;
+export declare const privateButtonLinkPressedForegroundNight = "rgb(209, 247, 255)" as const;
+export declare const privateButtonSecondaryOnDarkPressedBackgroundDay = "rgb(1, 9, 19)" as const;
+export declare const privateButtonPrimaryOnLightDisabledBackgroundNight = "rgb(11, 18, 29)" as const;
+export declare const privateButtonLinkPressedForegroundDay = "rgb(2, 77, 175)" as const;
+export declare const privateButtonDestructiveNormalBackgroundNight = "rgb(36, 51, 70)" as const;
+export declare const privateButtonPrimaryNormalBackgroundNight = "rgb(2, 77, 175)" as const;
+export declare const privateButtonDestructivePressedBackgroundNight = "rgb(255, 100, 156)" as const;
+export declare const privateButtonPrimaryPressedBackgroundNight = "rgb(5, 65, 132)" as const;
+export declare const privateButtonPrimaryOnLightDisabledBackgroundDay = "rgb(224, 228, 233)" as const;
+export declare const privateButtonPrimaryOnDarkDisabledBackgroundNight = "rgb(11, 18, 29)" as const;
+export declare const privateButtonPrimaryNormalBackgroundDay = "rgb(5, 32, 60)" as const;
+export declare const privateButtonDestructiveNormalBackgroundDay = "rgb(224, 228, 233)" as const;
+export declare const privateButtonLinkOnDarkPressedForegroundNight = "rgba(255, 255, 255, 0.5)" as const;
+export declare const privateButtonDestructivePressedBackgroundDay = "rgb(231, 8, 102)" as const;
+export declare const privateButtonPrimaryPressedBackgroundDay = "rgb(21, 70, 121)" as const;
+export declare const privateButtonPrimaryOnDarkDisabledBackgroundDay = "rgb(224, 228, 233)" as const;
+export declare const privateButtonLinkOnDarkNormalForegroundNight = "rgb(255, 255, 255)" as const;
+export declare const privateButtonDisabledBackgroundNight = "rgb(11, 18, 29)" as const;
+export declare const privateButtonSecondaryOnDarkNormalBackgroundNight = "rgba(255, 255, 255, 0.1)" as const;
+export declare const privateButtonSecondaryOnDarkDisabledForegroundNight = "rgba(255, 255, 255, 0.2)" as const;
+export declare const privateButtonLinkOnDarkPressedForegroundDay = "rgba(255, 255, 255, 0.5)" as const;
+export declare const privateButtonPrimaryOnDarkPressedBackgroundNight = "rgb(193, 199, 207)" as const;
+export declare const privateButtonLinkOnDarkNormalForegroundDay = "rgb(255, 255, 255)" as const;
+export declare const privateButtonDisabledBackgroundDay = "rgb(224, 228, 233)" as const;
+export declare const privateButtonSecondaryOnDarkNormalBackgroundDay = "rgba(255, 255, 255, 0.1)" as const;
+export declare const privateButtonSecondaryOnDarkDisabledForegroundDay = "rgba(255, 255, 255, 0.2)" as const;
+export declare const privateButtonPrimaryOnDarkPressedBackgroundDay = "rgb(193, 199, 207)" as const;
+export declare const privateButtonPrimaryOnLightPressedBackgroundNight = "rgb(5, 65, 132)" as const;
+export declare const privateButtonFeaturedNormalBackgroundNight = "rgb(132, 233, 255)" as const;
+export declare const privateBarTrackDefaultDay = "rgb(224, 228, 233)" as const;
+export declare const privateBarTrackDefaultNight = "rgb(36, 51, 70)" as const;
+export declare const privateBarTrackOnContrastDay = "rgb(255, 255, 255)" as const;
+export declare const privateBarTrackOnContrastNight = "rgb(36, 51, 70)" as const;
+export declare const privateSkeletonShimmerStartEndDay = "rgba(255, 255, 255, 0)" as const;
+export declare const privateSkeletonShimmerStartEndNight = "rgba(0, 0, 0, 0)" as const;
+export declare const privateSkeletonShimmerCenterDay = "rgba(255, 255, 255, 0.6)" as const;
+export declare const privateSkeletonShimmerCenterNight = "rgba(0, 0, 0, 0.2)" as const;
+export declare const privateSponsoredBannerBackgroundDay = "rgb(239, 243, 248)" as const;
+export declare const privateSponsoredBannerBackgroundNight = "rgb(36, 51, 70)" as const;
+export declare const privateNavigationTabHoverDay = "rgb(21, 70, 121)" as const;
+export declare const privateNavigationTabHoverNight = "rgb(209, 247, 255)" as const;
+export declare const privateNavigationTabOutlineDay = "rgb(193, 199, 207)" as const;
+export declare const privateNavigationTabOutlineNight = "rgb(255, 255, 255)" as const;
+export declare const privateCardButtonContainedFillDay = "rgba(255, 255, 255, 0.8)" as const;
+export declare const privateCardButtonContainedFillNight = "rgba(0, 0, 0, 0.8)" as const;
+export declare const durationXs = "50ms" as const;
+export declare const durationSm = "200ms" as const;
+export declare const durationBase = "400ms" as const;
+export declare const borderRadiusXs = ".25rem" as const;
+export declare const borderRadiusSm = ".5rem" as const;
+export declare const borderRadiusMd = ".75rem" as const;
+export declare const borderRadiusLg = "1.5rem" as const;
+export declare const borderRadiusXl = "2.5rem" as const;
+export declare const borderSizeSm = "1px" as const;
+export declare const borderSizeLg = "2px" as const;
+export declare const borderSizeXl = "3px" as const;
+export declare const boxShadowSm = "0px 1px 3px 0px rgba(37,32,31,.3)" as const;
+export declare const boxShadowLg = "0px 4px 14px 0px rgba(37,32,31,.25)" as const;
+export declare const boxShadowXl = "0px 12px 50px 0px rgba(37,32,31,.25)" as const;
+export declare const breakpointQueryTabletOnly = "(min-width: 32.0625rem) and (max-width: 64rem)" as const;
+export declare const breakpointSmallTabletMargin = "1.5rem" as const;
+export declare const breakpointQuerySmallTablet = "(max-width: 48rem)" as const;
+export declare const breakpointSmallMobileMargin = "1rem" as const;
+export declare const breakpointQueryMobile = "(max-width: 32rem)" as const;
+export declare const breakpointQuerySmallTabletOnly = "(min-width: 32.0625rem) and (max-width: 48rem)" as const;
+export declare const breakpointSmallMobile = "22.4375rem" as const;
+export declare const breakpointQueryAboveMobile = "(min-width: 32.0625rem)" as const;
+export declare const breakpointQueryTablet = "(max-width: 64rem)" as const;
+export declare const breakpointSmallTablet = "48rem" as const;
+export declare const breakpointQueryAboveTablet = "(min-width: 64.0625rem)" as const;
+export declare const breakpointDesktop = "80rem" as const;
+export declare const breakpointMobile = "32rem" as const;
+export declare const breakpointTablet = "64rem" as const;
+export declare const breakpointQueryDesktopOnly = "(min-width: 64.0625rem) and (max-width: 80rem)" as const;
+export declare const breakpointQueryAboveDesktop = "(min-width: 80.0625rem)" as const;
+export declare const breakpointTabletMargin = "2rem" as const;
+export declare const breakpointMobileMargin = "1.5rem" as const;
+export declare const breakpointQuerySmallMobile = "(max-width: 22.4375rem)" as const;
+export declare const breakpointDesktopMargin = "3rem" as const;
+export declare const buttonHeight = "2.25rem" as const;
+export declare const buttonBorderRadius = ".5rem" as const;
+export declare const buttonLargeHeight = "3rem" as const;
+export declare const privateButtonLineHeight = "1.5rem" as const;
+export declare const privateCardButtonContainedColor = "rgba(255, 255, 255, 0.8)" as const;
+export declare const calendarDaySpacing = ".5rem" as const;
+export declare const calendarDaySize = "2.25rem" as const;
+export declare const cardBackgroundColor = "rgb(255, 255, 255)" as const;
+export declare const cardPadding = "1rem" as const;
+export declare const primaryGradient = "#0062E3" as const;
+export declare const flareHeightDesktop = "1.5rem" as const;
+export declare const flareHeightMobile = "1rem" as const;
+export declare const flareCornerRadius = "1.3125rem" as const;
+export declare const inputLargeHeight = "3rem" as const;
+export declare const inputBorder = "solid .0625rem #C1C7CF" as const;
+export declare const selectBorderRadius = ".25rem" as const;
+export declare const inputBackground = "#ffffff" as const;
+export declare const inputBorderWidth = ".0625rem" as const;
+export declare const labelDisabledColor = "rgba(0, 0, 0, 0.2)" as const;
+export declare const inputBorderRadius = ".25rem" as const;
+export declare const selectDisabledBorderColor = "rgb(224, 228, 233)" as const;
+export declare const inputDisabledColor = "rgba(0, 0, 0, 0.2)" as const;
+export declare const inputPaddingX = ".5rem" as const;
+export declare const inputPaddingY = ".5rem" as const;
+export declare const selectBorderWidth = ".0625rem" as const;
+export declare const formValidationColor = "rgb(231, 8, 102)" as const;
+export declare const inputHeight = "2.25rem" as const;
+export declare const inputDisabledBorderColor = "rgb(239, 243, 248)" as const;
+export declare const horizontalNavBarSelectedColor = "rgb(0, 98, 227)" as const;
+export declare const iconSizeSm = "1rem" as const;
+export declare const iconSizeLg = "1.5rem" as const;
+export declare const modalBackgroundColor = "rgb(255, 255, 255)" as const;
+export declare const modalInitialOpacity = "0" as const;
+export declare const modalOpacity = "1" as const;
+export declare const modalMaxWidth = "32rem" as const;
+export declare const modalWideMaxWidth = "64rem" as const;
+export declare const modalContentPadding = "1rem" as const;
+export declare const bannerAlertHeaderExpandableHoverBackgroundColor = "rgb(239, 243, 248)" as const;
+export declare const bannerAlertHeaderExpandableActiveBackgroundColor = "rgb(239, 243, 248)" as const;
+export declare const bannerAlertPrimaryColor = "rgb(0, 98, 227)" as const;
+export declare const bannerAlertSuccessColor = "rgb(12, 131, 138)" as const;
+export declare const bannerAlertWarnColor = "rgb(245, 93, 66)" as const;
+export declare const bannerAlertErrorColor = "rgb(231, 8, 102)" as const;
+export declare const bannerAlertNeutralColor = "rgb(193, 199, 207)" as const;
+export declare const panelBorderColor = "rgb(193, 199, 207)" as const;
+export declare const privateSliderSelectedDay = "rgb(21, 70, 121)" as const;
+export declare const onePixelRem = ".0625rem" as const;
+export declare const spacingNone = "0" as const;
+export declare const spacingIconText = ".5rem" as const;
+export declare const fontWeightBook = "400" as const;
+export declare const lineHeightXlTight = "1.75rem" as const;
+export declare const lineHeightXxxxxl = "4.5rem" as const;
+export declare const lineHeightXxxl = "3rem" as const;
+export declare const lineHeightXl = "2rem" as const;
+export declare const fontSizeXl = "1.5rem" as const;
+export declare const fontFamilyLarken = "'Larken', 'Noto Sans', 'Noto Sans JP', sans-serif" as const;
+export declare const fontSizeXxxxxl = "4rem" as const;
+export declare const fontSizeXxxl = "2.5rem" as const;
+export declare const lineHeightSm = "1.25rem" as const;
+export declare const lineHeightLgTight = "1.5rem" as const;
+export declare const lineHeightLg = "1.75rem" as const;
+export declare const fontWeightBlack = "900" as const;
+export declare const lineHeightXs = "1rem" as const;
+export declare const fontSizeSm = ".875rem" as const;
+export declare const fontSizeXs = ".75rem" as const;
+export declare const letterSpacingTight = "-0.02em" as const;
+export declare const fontSizeLg = "1.25rem" as const;
+export declare const fontSizeRoot = "100%" as const;
+export declare const fontSize6Xl = "4.75rem" as const;
+export declare const fontFamilyBase = "'Skyscanner Relative', -apple-system, BlinkMacSystemFont, 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif" as const;
+export declare const fontWeightBold = "700" as const;
+export declare const fontColorBase = "rgb(22, 22, 22)" as const;
+export declare const lineHeightBaseTight = "1.25rem" as const;
+export declare const fontSize7Xl = "6rem" as const;
+export declare const lineHeight6Xl = "5.25rem" as const;
+export declare const fontSize8Xl = "7.5rem" as const;
+export declare const fontSizeXxl = "2rem" as const;
+export declare const fontSizeXxxxl = "3rem" as const;
+export declare const lineHeightBase = "1.5rem" as const;
+export declare const lineHeight7Xl = "6rem" as const;
+export declare const fontSizeBase = "1rem" as const;
+export declare const fontWeightLight = "300" as const;
+export declare const lineHeight8Xl = "7.5rem" as const;
+export declare const lineHeightXxl = "2.5rem" as const;
+export declare const lineHeightXxxxl = "3.5rem" as const;
+export declare const zindexAutosuggest = "900" as const;
+export declare const zindexPopover = "900" as const;
+export declare const zindexTooltip = "900" as const;
+export declare const zindexScrim = "1000" as const;
+export declare const zindexModal = "1100" as const;
+export declare const zindexDrawer = "1100" as const;
+export declare const animations = {
+durationXs,
+durationSm,
+durationBase,
+} as const;
+export declare const badgeColors = {
+privateBadgeBackgroundNormalDay,
+privateBadgeBackgroundNormalNight,
+} as const;
+export declare const borders = {
+borderSizeSm,
+borderSizeLg,
+borderSizeXl,
+} as const;
+export declare const boxShadows = {
+boxShadowSm,
+boxShadowLg,
+boxShadowXl,
+} as const;
+export declare const breakpoints = {
+breakpointQueryTabletOnly,
+breakpointSmallTabletMargin,
+breakpointQuerySmallTablet,
+breakpointSmallMobileMargin,
+breakpointQueryMobile,
+breakpointQuerySmallTabletOnly,
+breakpointSmallMobile,
+breakpointQueryAboveMobile,
+breakpointQueryTablet,
+breakpointSmallTablet,
+breakpointQueryAboveTablet,
+breakpointDesktop,
+breakpointMobile,
+breakpointTablet,
+breakpointQueryDesktopOnly,
+breakpointQueryAboveDesktop,
+breakpointTabletMargin,
+breakpointMobileMargin,
+breakpointQuerySmallMobile,
+breakpointDesktopMargin,
+} as const;
+export declare const buttonColors = {
+privateButtonSecondaryPressedBackgroundNight,
+privateButtonLinkNormalForegroundNight,
+privateButtonLinkOnDarkDisabledForegroundNight,
+privateButtonSecondaryNormalBackgroundNight,
+privateButtonPrimaryOnLightPressedBackgroundDay,
+privateButtonFeaturedNormalBackgroundDay,
+privateButtonFeaturedPressedBackgroundNight,
+privateButtonPrimaryOnLightNormalBackgroundNight,
+privateButtonSecondaryPressedBackgroundDay,
+privateButtonLinkNormalForegroundDay,
+privateButtonPrimaryOnLightDisabledForegroundNight,
+privateButtonSecondaryOnDarkDisabledBackgroundNight,
+privateButtonDestructiveNormalForegroundNight,
+privateButtonLinkOnDarkDisabledForegroundDay,
+privateButtonSecondaryNormalBackgroundDay,
+privateButtonPrimaryOnLightNormalBackgroundDay,
+privateButtonPrimaryOnLightDisabledForegroundDay,
+privateButtonPrimaryOnDarkNormalBackgroundNight,
+privateButtonFeaturedPressedBackgroundDay,
+privateButtonPrimaryOnDarkDisabledForegroundNight,
+privateButtonSecondaryOnDarkDisabledBackgroundDay,
+privateButtonSecondaryOnDarkPressedBackgroundNight,
+privateButtonDestructiveNormalForegroundDay,
+privateButtonPrimaryOnDarkNormalBackgroundDay,
+privateButtonPrimaryOnDarkDisabledForegroundDay,
+privateButtonLinkPressedForegroundNight,
+privateButtonSecondaryOnDarkPressedBackgroundDay,
+privateButtonPrimaryOnLightDisabledBackgroundNight,
+privateButtonLinkPressedForegroundDay,
+privateButtonDestructiveNormalBackgroundNight,
+privateButtonPrimaryNormalBackgroundNight,
+privateButtonDestructivePressedBackgroundNight,
+privateButtonPrimaryPressedBackgroundNight,
+privateButtonPrimaryOnLightDisabledBackgroundDay,
+privateButtonPrimaryOnDarkDisabledBackgroundNight,
+privateButtonPrimaryNormalBackgroundDay,
+privateButtonDestructiveNormalBackgroundDay,
+privateButtonLinkOnDarkPressedForegroundNight,
+privateButtonDestructivePressedBackgroundDay,
+privateButtonPrimaryPressedBackgroundDay,
+privateButtonPrimaryOnDarkDisabledBackgroundDay,
+privateButtonLinkOnDarkNormalForegroundNight,
+privateButtonDisabledBackgroundNight,
+privateButtonSecondaryOnDarkNormalBackgroundNight,
+privateButtonSecondaryOnDarkDisabledForegroundNight,
+privateButtonLinkOnDarkPressedForegroundDay,
+privateButtonPrimaryOnDarkPressedBackgroundNight,
+privateButtonLinkOnDarkNormalForegroundDay,
+privateButtonDisabledBackgroundDay,
+privateButtonSecondaryOnDarkNormalBackgroundDay,
+privateButtonSecondaryOnDarkDisabledForegroundDay,
+privateButtonPrimaryOnDarkPressedBackgroundDay,
+privateButtonPrimaryOnLightPressedBackgroundNight,
+privateButtonFeaturedNormalBackgroundNight,
+} as const;
+export declare const buttons = {
+buttonHeight,
+buttonBorderRadius,
+buttonLargeHeight,
+privateButtonLineHeight,
+privateCardButtonContainedColor,
+} as const;
+export declare const calendar = {
+calendarDaySpacing,
+calendarDaySize,
+} as const;
+export declare const canvasColors = {
+canvasDay,
+canvasNight,
+canvasContrastDay,
+canvasContrastNight,
+} as const;
+export declare const cardButtonColors = {
+privateCardButtonContainedFillDay,
+privateCardButtonContainedFillNight,
+} as const;
+export declare const cards = {
+cardBackgroundColor,
+cardPadding,
+} as const;
+export declare const chipColors = {
+privateChipOnDarkPressedStrokeDay,
+privateChipOnDarkPressedStrokeNight,
+privateChipOnDarkOnBackgroundDay,
+privateChipOnDarkOnBackgroundNight,
+privateChipDisabledBackgroundDay,
+privateChipDisabledBackgroundNight,
+privateChipOnDarkOnDismissIconDay,
+privateChipOnDarkOnDismissIconNight,
+} as const;
+export declare const colors = {
+colorSkyBlueShade03,
+colorPrimaryGradientLight,
+colorErfoud,
+colorValensole,
+colorMonteverde,
+colorSkyGrayTint01,
+colorBlack,
+colorSkyGrayTint02,
+colorSagano,
+colorSkyGrayTint03,
+colorSkyGrayTint04,
+colorHarbour,
+colorSkyGrayTint05,
+colorSkyGrayTint06,
+colorBlackTint01,
+colorSkyGrayTint07,
+colorBlackTint02,
+colorAbisko,
+colorBlackTint03,
+colorBlackTint04,
+colorWhite,
+colorBlackTint05,
+colorBlackTint06,
+colorPanjin,
+colorKolkata,
+colorGlencoe,
+colorTochigi,
+colorPetra,
+colorSkyBlue,
+colorBagan,
+colorHillier,
+colorSkyBlueTint01,
+colorSkyBlueTint02,
+colorBunol,
+colorSkyBlueTint03,
+colorSkyGray,
+colorNara,
+colorSkyBlueShade01,
+colorSkyBlueShade02,
+lineDarkColor,
+backgroundAlternativeSecondaryDarkColor,
+backgroundLightColor,
+backgroundDarkColor,
+backgroundAlternativeSecondaryLightColor,
+colorSystemGreen,
+lineLightColor,
+backgroundTertiaryDarkColor,
+backgroundSecondaryDarkColor,
+backgroundAlternativeLightColor,
+primaryDarkColor,
+colorSystemRed,
+backgroundAlternativeDarkColor,
+primaryLightColor,
+backgroundTertiaryLightColor,
+backgroundSecondaryLightColor,
+primaryGradient,
+} as const;
+export declare const coreColors = {
+coreEcoDay,
+coreEcoNight,
+coreAccentDay,
+coreAccentNight,
+corePrimaryDay,
+corePrimaryNight,
+} as const;
+export declare const flare = {
+flareHeightDesktop,
+flareHeightMobile,
+flareCornerRadius,
+} as const;
+export declare const fontWeights = {
+fontWeightBook,
+fontWeightBlack,
+fontWeightBold,
+fontWeightLight,
+} as const;
+export declare const forms = {
+inputLargeHeight,
+inputBorder,
+selectBorderRadius,
+inputBackground,
+inputBorderWidth,
+labelDisabledColor,
+inputBorderRadius,
+selectDisabledBorderColor,
+inputDisabledColor,
+inputPaddingX,
+inputPaddingY,
+selectBorderWidth,
+formValidationColor,
+inputHeight,
+inputDisabledBorderColor,
+} as const;
+export declare const horizontalNav = {
+horizontalNavBarSelectedColor,
+} as const;
+export declare const icons = {
+iconSizeSm,
+iconSizeLg,
+} as const;
+export declare const infoBannerColors = {
+privateInfoBannerDefaultDay,
+privateInfoBannerOnContrastDay,
+privateInfoBannerSuccessDay,
+privateInfoBannerErrorNight,
+privateInfoBannerInfoNight,
+privateInfoBannerErrorDay,
+privateInfoBannerInfoDay,
+privateInfoBannerWarningNight,
+privateInfoBannerWarningDay,
+privateInfoBannerDefaultNight,
+privateInfoBannerOnContrastNight,
+privateInfoBannerSuccessNight,
+} as const;
+export declare const letterSpacings = {
+letterSpacingTight,
+} as const;
+export declare const lineColors = {
+lineOnDarkNight,
+lineOnDarkDay,
+lineNight,
+lineDay,
+} as const;
+export declare const mapMarkerColors = {
+privateMapPreviousSelectionDay,
+privateMapClusterPinNight,
+privateMapClusterPinPreviousSelectionNight,
+privateMapClusterPinDay,
+privateMapClusterPinPreviousSelectionDay,
+privateMapMarkerViewedForegroundNight,
+privateMapMarkerViewedForegroundDay,
+privateMapPoiPinNight,
+privateMapPreviousSelectionNight,
+privateMapPoiPinDay,
+} as const;
+export declare const marcommsColors = {
+marcommsBerry,
+marcommsCharcoal,
+marcommsOrange,
+marcommsGrey40,
+marcommsBlueBright,
+marcommsYellow,
+marcommsGreenBright,
+marcommsPink,
+marcommsSkyBlue,
+marcommsDarkSky,
+marcommsPurple,
+marcommsWhite,
+marcommsYellowMuted,
+marcommsPinkMuted,
+marcommsGreenMuted,
+marcommsBlueMuted,
+marcommsEcoGreen,
+} as const;
+export declare const modals = {
+modalBackgroundColor,
+modalInitialOpacity,
+modalOpacity,
+modalMaxWidth,
+modalWideMaxWidth,
+modalContentPadding,
+} as const;
+export declare const navigationTabColors = {
+privateNavigationTabHoverDay,
+privateNavigationTabHoverNight,
+privateNavigationTabOutlineDay,
+privateNavigationTabOutlineNight,
+} as const;
+export declare const notifications = {
+bannerAlertHeaderExpandableHoverBackgroundColor,
+bannerAlertHeaderExpandableActiveBackgroundColor,
+bannerAlertPrimaryColor,
+bannerAlertSuccessColor,
+bannerAlertWarnColor,
+bannerAlertErrorColor,
+bannerAlertNeutralColor,
+} as const;
+export declare const panels = {
+panelBorderColor,
+} as const;
+export declare const radii = {
+borderRadiusXs,
+borderRadiusSm,
+borderRadiusMd,
+borderRadiusLg,
+borderRadiusXl,
+} as const;
+export declare const ratingBarColors = {
+privateBarTrackDefaultDay,
+privateBarTrackDefaultNight,
+privateBarTrackOnContrastDay,
+privateBarTrackOnContrastNight,
+} as const;
+export declare const scrimColors = {
+scrimDay,
+scrimNight,
+} as const;
+export declare const skeletonColors = {
+privateSkeletonShimmerStartEndDay,
+privateSkeletonShimmerStartEndNight,
+privateSkeletonShimmerCenterDay,
+privateSkeletonShimmerCenterNight,
+} as const;
+export declare const sliderColors = {
+privateSliderSelectedDay,
+} as const;
+export declare const spacings = {
+onePixelRem,
+spacingNone,
+spacingIconText,
+} as const;
+export declare const sponsoredBannerColors = {
+privateSponsoredBannerBackgroundDay,
+privateSponsoredBannerBackgroundNight,
+} as const;
+export declare const statusColors = {
+statusSuccessSpotNight,
+statusSuccessFillNight,
+statusSuccessSpotDay,
+statusSuccessFillDay,
+statusDangerSpotNight,
+statusWarningSpotNight,
+statusDangerFillNight,
+statusWarningFillNight,
+statusDangerSpotDay,
+statusWarningSpotDay,
+statusDangerFillDay,
+statusWarningFillDay,
+} as const;
+export declare const surfaceColors = {
+surfaceHeroNight,
+surfaceDefaultNight,
+surfaceHeroDay,
+surfaceDefaultDay,
+surfaceContrastNight,
+surfaceContrastDay,
+surfaceElevatedNight,
+surfaceSubtleNight,
+surfaceElevatedDay,
+surfaceHighlightNight,
+surfaceSubtleDay,
+surfaceHighlightDay,
+} as const;
+export declare const textColors = {
+textOnDarkDay,
+textOnLightNight,
+textDisabledOnDarkDay,
+textOnLightDay,
+textLinkNight,
+textHeroNight,
+textDisabledNight,
+textPrimaryDarkColor,
+textLinkDay,
+textHeroDay,
+textQuaternaryDarkColor,
+textPrimaryNight,
+textQuaternaryLightColor,
+textDisabledDay,
+textPrimaryLightColor,
+textPrimaryDay,
+textErrorNight,
+textErrorDay,
+textTertiaryDarkColor,
+textSecondaryDarkColor,
+textPrimaryInverseNight,
+textSecondaryNight,
+textOnDarkNight,
+textSecondaryLightColor,
+textPrimaryInverseDay,
+textDisabledOnDarkNight,
+textTertiaryLightColor,
+textSecondaryDay,
+backgroundElevation03DarkColor,
+backgroundElevation02DarkColor,
+backgroundElevation01DarkColor,
+backgroundElevation01LightColor,
+backgroundElevation02LightColor,
+backgroundElevation03LightColor,
+fontColorBase,
+} as const;
+export declare const typesettings = {
+lineHeightXlTight,
+lineHeightXxxxxl,
+lineHeightXxxl,
+lineHeightXl,
+fontSizeXl,
+fontFamilyLarken,
+fontSizeXxxxxl,
+fontSizeXxxl,
+lineHeightSm,
+lineHeightLgTight,
+lineHeightLg,
+lineHeightXs,
+fontSizeSm,
+fontSizeXs,
+fontSizeLg,
+fontSizeRoot,
+fontSize6Xl,
+fontFamilyBase,
+lineHeightBaseTight,
+fontSize7Xl,
+lineHeight6Xl,
+fontSize8Xl,
+fontSizeXxl,
+fontSizeXxxxl,
+lineHeightBase,
+lineHeight7Xl,
+fontSizeBase,
+lineHeight8Xl,
+lineHeightXxl,
+lineHeightXxxxl,
+} as const;
+export declare const zIndices = {
+zindexAutosuggest,
+zindexPopover,
+zindexTooltip,
+zindexScrim,
+zindexModal,
+zindexDrawer,
+} as const;